requirements.txt
Django==1.11.4
djangorestframework==3.6.3
PyJWT==1.5.2
pytz==2017.2
